refactor: big and beautiful
