+++ internal/pg/git.go
@@ -3,9 +3,10 @@ package pg
 import (
 	"database/sql"
 
-	"github.com/dsc-sgu/mm-backend/internal/git"
 	"github.com/google/uuid"
 	"go.uber.org/zap"
+
+	"github.com/dsc-sgu/mm-backend/internal/git"
 )
 
 const (
@@ -13,6 +14,7 @@ const (
 		INSERT INTO ssh_keys (owner_id, name, key, fingerprint, created_at)
 		VALUES (:owner_id, :name, :key, :fingerprint, :created_at)
 	`
+
 	deleteSshKeySQL = `
 		DELETE FROM ssh_keys
 		WHERE owner_id = $1 AND fingerprint = $2
@@ -42,4 +44,17 @@ func (r *PGRepo) DeleteSshKey(ownerId uuid.UUID, fingerprint string) error {
 	}
 
 	return nil
+}
+
+func GetTask(name string) (uuid.UUID, error) {
+	return uuid.Parse("e7cf6012-1348-434b-9d54-bd89c9e6e95e")
+}
+
+// realise via db
+func GetParticipant(fingerprint string) (uuid.UUID, error) {
+	if fingerprint == "SHA256:AH71wflD7hbxs0bGhssvTy77dLoszYUWXkeK798ph04" {
+		return uuid.Parse("681ae49f-1f56-4632-bd1b-7ca3ab09a467")
+	} else {
+		return uuid.Parse("e7cf6012-1348-434b-9d54-bd89c9e6e95e")
+	}
 }
